return notification dto

+++ MochiApi/Dtos/Budget/BudgetDetailSummary.cs
@@ -2,8 +2,8 @@
 {
     public class BudgetDetailSummary : BudgetSummary
     {
-        public long RecommendedDailyExpense { get; set; }
-        public long RealDailyExpense { get; set; }
-        public long ExpectedExpense { get; set; }
+        public double RecommendedDailyExpense { get; set; }
+        public double RealDailyExpense { get; set; }
+        public double ExpectedExpense { get; set; }
     }
 }
